requirements.txt
streamlit
paddlepaddle==2.62
paddleocr=2.7.3
numpy<2.0
opencv-python-headless
pandas
Pillow
shapely